Move slider settings to separate component
